+++ src/card.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
-import Badge from './badge';
-import { color as ValidColor } from './types/color';
+
 import cn from './utilities/classnames';
 
+import Badge from './badge';
+import Icon from './icon';
+
+import { color as ValidColor } from './types/color';
+import { IconType } from './types/icons';
+
 export interface CardProps {
-  children?: React.ReactElement<any>;
+  children?: React.ReactNode;
   centered?: boolean;
   thin?: boolean;
   inline?: boolean;
@@ -17,6 +22,9 @@ export interface CardProps {
         color: ValidColor;
       }
     | React.ReactElement<any>;
+  icon?: IconType;
+  iconSize?: IconSizeType;
+  selected?: boolean;
 }
 
 export const Card: React.SFC<CardProps> = ({
@@ -24,27 +32,33 @@ export const Card: React.SFC<CardProps> = ({
   body,
   centered,
   children,
+  icon,
+  iconSize,
   inline,
   thin,
   title,
   className,
+  selected,
   ...attributes,
 }) => {
-
-  const titleStyle = {
-    width: '100%',
-  } as React.CSSProperties;
 
   return (
     <div
       className={cn('card', className, {
         'is-centered': centered,
+        'is-inline': inline,
+        'is-selected': selected,
         'is-thin': thin,
       })}
       {...attributes}
     >
       {badge && <Badge {...badge} />}
-      {title && <h2 style={titleStyle}>{title}</h2>}
+      {icon &&
+        <div className="card-icon">
+          <Icon size={iconSize} type={icon} />
+        </div>
+      }
+      {title && <h2 className={cn({ 'card-title': inline })}>{title}</h2>}
       {body && <p>{body}</p>}
       {children}
     </div>